fix(segment): handle race condition on loading segment integration (#586)
* fix(segment): handle race condition on loading segment integration

analytics.register is async, thus we need to wait for it to finish
before we can call e.g. analytics.page(). Prior to this, the loaded
callback would be call before the integration was loaded.

This was a result of issues raised by users [in
ZenDesk](https://posthoghelp.zendesk.com/agent/tickets/2082) and [in
Slack](https://posthogusers.slack.com/archives/C01GLBKHKQT/p1678950622505599).

* add back initPromises

* fix test, bind instance

* wip

* wip

* wip

* Add core-js explicitly

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* drop version change